Motor 1 runs properly, python interface can be used to control motors. Next: test rest of motors and figure out reversing

+++ src/structures.h
@@ -14,5 +14,5 @@
 
 typedef struct{
     int8_t payload[6];
-    int new;
+    uint8_t new;
 }MotorCmdFrame;
